Finance.jl Documentation

Overview

Below is a collection of functions which are useful to find signals in time series data.

sigcumsum: This function looks for imbalance in tics.

isConvertible: This function determines if one type can be converted to another.

tic_diff1: As tics are irregular, this function provides a way to compute the numerical derivative with respect to an irregular time series.

tic_diff2: This function finds the numerical second derivative with respect to irregular time series data.

ema: Computes the Exponential Moving Average of a time series.

ema_std: Computes the standard deviation of an Exponential Moving Average of a time series.

ema_stat: Computes the Exponential Moving Average along with the associated moving standard deviation, relative skewness, relative kurtosis.

ewt_mean Computes a exponentially decayed temporal weighted moving average.

pow_n: Computes the power of a value to a positive integer power using repeated squaring.

Signals

Finance.sig_cumsum — Function

sig_cumsum(t, x, w, h; chk_inp=false)

Return a tuple of two vectors: tics, signals.

The signals are the collections of all deviations from a running mean (with window length w) of the series. The threshold of the deviation is h.

Deviation is determined by:

$S_t^+ = {\rm max}(0, S_{t-1} + x_t - E[x_{t-1}]; S^+_0 = 0$

$S_t^- = {\rm min}(0, S_{t-1} + x_t - E[x_{t-1}]; S^-_0 = 0$

$S_t^{\hphantom{+}} = {\rm max}(S^+_t, -S^-_t)$

Collect all $t, S_t$ where $h \ge S_t$.

Type Constraints

S <: Real

T <: Real

Arguments

t :: AbstractVector{S} – The tic series to examine.

x :: AbstractVector{T} – The series to examine.

w :: Int – The width of the moving average.

h :: T – The threshold for the deviation to register.

Keyword Arguments

chk_inp=false :: Bool – Check the input contract?

Input Contract

The inputs are assumed to satisfy the constraints below.

Constraint	Description

|x| $\ge$ 2	The length of x is at least $2$.

|x| = |t|	The length of x is equal to the length of t.

w > 1	The window length is greater than $1$.

h > 0	The deviation threshold is greater than 0.

$\forall i, t_{i+1} > t_{i}$	The times are increasing; consequently, we have a 1-1 map from t to x.

Output Components

td :: AbstractVector{S} – Values of t where deviations occurred.

xd :: AbstractVector{T} – Values of x where deviations occurred.

Output Contract

|td| = |xd|

Return

(td, xd) :: Tuple{AbstractVector{S}, AbstractVector{T}}

Moving Averages

Finance.ema — Function

ema(x,m; h=div(m,2))

Compute the Exponential Moving Average of the sequence x.

Type Constraints

T <: Real

Arguments

x :: AbstractVector{T} – The series to work with.

m :: Int – The width of the decay window.

Keyword Arguments

h=div(m,2) :: Int – The exponential decay half-life.

Input Contract

The inputs are assumed to satisfy the constraints below.

Constraint	Description

m > 1	Averaging window length is greater than $1$.

h > 1	Exponential half-life is greater than $1$.

Output

ema :: AbstractVector{T} – The exponential moving average of x.

Output Contract

|x| = |ema|

Return

ema::AbstractVector{T}

Finance.ema_std — Function

ema_std(x, m; h=div(m,2), init_sig=nothing)

Compute the Moving Exponential Standard Deviation of the sequence x. By default, the initial std is taken to be the standard deviation of the first window (of length m). However, a user specified value may be used instead.

Type Constraints

T <: Real

Arguments

x :: AbstractVector{T} – The series to work with.

m :: Int – The width of the decay window.

Keyword Arguments

h=div(m,2) :: Int – The exponential decay half-life.

init_sig=nothing :: Union{Nothing, T} – An optional user supplied initial standard deviation for the start of the series.

Input Contract

The inputs are assumed to satisfy the constraints below.

Constraint	Description

m > 1	Averaging window length is greater than $1$.

h > 1	Exponential half-life is greater than $1$.

|x| > 1	The length of the series is greater than $1$.

init_sig $\ge 0$	User supplied starting $\sigma$ should be $\ge 0$.

Output

stda :: AbstractVector{T} – The moving exponential standard deviation of x.

Output Contract

|x| = |stda|

Return

stda::Vector{T}

Finance.ema_stats — Function

ema_stats(x, m; h=div(m,2), init_sig=nothing)

Compute the Moving Exponential Stats of the sequence x: ema, ema_std, ema_rel_skew, ema_rel_kurtosis.

The recursive formulas for the moving statistics as well as the adjustments necessary to render the estimates unbiased come from the paper: exponential_moving_average.pdf.

Returns these stats as a matrix with four columns, each representing the stats above in the order listed.

Type Constraints

T <: Real

Arguments

x :: AbstractVector{T} – The series to work with.

m :: Int – The width of the decay window.

Keyword Arguments

h=div(m,2) :: Int – The exponential decay half-life.

init_sig=nothing:: Union{T, Nothing} – An optional user supplied initial standard deviation for the start of the series.

Input Contract

The inputs are assumed to satisfy the constraints below:

Constraint	Description

m > 1	Averaging window length is greater than $1$.

h > 1	Exponential half-life is greater than $1$.

|x| > 3	The length of the series is greater than $3$.

Output

stat :: Matrix{T} – A matrix of EMA stats: ema, ema_std, ema_rel_skew, ema_rel_kurtosis.

Output Contract

|stat| = (N, 4)

Return

stat::Matrix{T}

Finance.std — Function

std(x)

Compute the "sample" standard deviation of a series, x.

Type Constraints

T <: Real

Arguments

x :: AbstractVector{T} – The series to work with.

Input Contract

The inputs are assumed to satisfy the constraints below.

Constraint	Description

|x| > 1	The length of the series is greater than $1$.

Return

std::T – The sample standard deviation.

Utilities

Finance.isConvertible — Function

isConvertible(S, T)

Boolean function which returns true if a value of type S can be converted to a value of type T.

Type Constraints

S <: Real

T <: Real

Arguments

::Type{S} – A numeric type.

::Type{T} – A numeric type.

Return

::Bool

Finance.tic_diff1 — Method

tic_diff1(t, x; chk_inp=false)

Compute the numerical derivative of a function represented by x with respect to t when the values in t are possibly irregular.

Type Constraints

S <: Real

T <: Real

Arguments

t :: AbstractVector{T} – A vector of times.

x :: AbstractVector{S} – A vector of values.

Keyword Arguments

chk_inp=false :: Bool – Check the input contract?

Input Contract

The inputs are assumed to satisfy the constraints below.

Constraint	Description

|t| = |x|	The length of the time and data vectors match.

T => S	Type T can be converted to type S.

$\forall i, t_{i+1} > t_i$	The times are increasing; consequently, we have a 1-1 map from t to x.

Return

:: AbstractVector{S}

Finance.tic_diff2 — Method

tic_diff2(t, x; chk_inp=false)

Compute the numerical second derivative of a function represented by x with respect to t when the values in t are possibly irregular.

Type Constraints

T <: Real

S <: Real

Arguments

t :: AbstractVector{T} – A vector of times.

x :: AbstractVector{S} – A vector of values.

Keyword Arguments

chk_inp=false :: Bool – Check the input contract?

Input Contract

The inputs are assumed to satisfy the constraints below.

Constraint	Description

|t| = |x|	The length of the time and data vectors match.

T => S	Type T can be converted to type S.

$\forall i, t_{i+1} > t_i$	The times are increasing; consequently, we have a 1-1 map from t to x.

Return

:: AbstractVector{S}

Finance.pow_n — Method

pow_n(x, n)

Fast (non-negative) integer powers: $x^n$. Uses repeated squaring in combination with the bit vector representation of n.

Type Constraints

T <: Number

Arguments

x::T – The base value.

n::Int – The power.

Input Contract

$n \ge 0$

Return

::T – The Power Value.

Finance.pow_n — Method

pow_n(x, n, m)

Fast integer (non-negative) powers with modulus: $x^n \; {\rm mod } \; m$. Uses repeated squaring in combination with the bit vector representation of n.

The output will be of type T^* = typeof(promote(x, m)).

Type Constraints

T <: Real

S <: Real

Arguments

x::T – The base value.

n::Int – The power.

m::S – The modulus.

Input Contract

$n \ge 0$

Return

$::T^*$ – The Power Value mod m.

Finance.entropy_index — Method

entropy_index(x; <keyword arguments>)

Computes a (Discounted) Binned Entropy Index. This is the ratio of entropy of the binned distribution of x against the entropy of the uniform distribution. The vector x is first capped by the lower and upper quantiles; then binned into n number of equal width bins. A distribution is formed from the bins and the entropy computed. If λ is not 1, then a discounted entropy is computed. This is an exponentially based discounting of the bin distribution based on their "freshness". In either event, the ratio of this entropy to the entropy of the corresponding uniform distribution (of n bins) is returned.

Type Constraints

T <: Real

F <: Float64

Arguments

x::AbstractVector{T} – Vector to process.

Keyword Arguments

n=10::Int – Exponential.

tol=1.0/(100 * n)::F – Error tolerance used with equivalency test of number to 0 or 1.

probs=[0.01, 0.99]::Vector{F} – Vector of quantile min and max.

λ=1.0::F – Discount value.

Input Contract

n > 2

0 < tol < 0.01

|probs| == 2

$0 < \lambda \le 1$

Return

::F – The (discounted) binned entropy index.

Finance.ewt_mean — Method

ewt_mean(ts, xs, b, lm)

Computes the moving (exponential decayed) temporal average of the data xs over windows of length b. Temporal averaging over a window means that the time stamps are differenced and we associate the difference of time points $t_{i}, t_{i+1}$, $\Delta_{i} = t_{i+1} - t_i$, with the data point $x_i$. The rationale: The data point $x_i$ has been around since $t_i$ until $t_{i+1}$, so it should be should weight it (in an un-normalized way) by this distance. The temporal decay will adjust the temporal weights by an exponential which puts more weight on recent data within the window. This is also done in an un-normalized way. Then the weights are normalized and data, xs, is averaged within the window.

NOTE: Set lm to 1.0 to just have temporal weighting without decay.

Parameters

ts::AbstractVector{Float64} – Data time stamps – ordered from smallest (oldest) to largest (newest).

xs::AbstractVector{Float64} – Data values associated with time stamps.

b::Int – The width of the window

lm::Float64 – The decay factor: 0.0 < lm <= 1.0

Input Contract

|ts| == |xs|

0 < b < |xs|

0.0 < lm <= 1.0

Return

::Vector{Float64} – A vector of length |xs| - b.

Index

Finance.ema

Finance.ema_stats

Finance.ema_std

Finance.entropy_index

Finance.ewt_mean

Finance.isConvertible

Finance.pow_n

Finance.pow_n

Finance.sig_cumsum

Finance.std

Finance.tic_diff1

Finance.tic_diff2